Fixed a couple more VS warnings.  Made common.c's initialize_demo() actually use w and h.

+++ demos/common/common.h
@@ -1,8 +1,10 @@
 #ifndef _COMMON_H__
 #define _COMMON_H__
 
+#include "SDL_gpu.h"
+
 void printRenderers(void);
 void printCurrentRenderer(void);
-GPU_Target* initialize_demo(int argc, char** argv, int w, int h);
+GPU_Target* initialize_demo(int argc, char** argv, Uint16 w, Uint16 h);
 
 #endif
